Break out debug and add -NUM actors
