+++ functions.php
@@ -357,13 +357,14 @@ new LWTV_Functions();
 require_once 'features/_main.php';     // General Features: This MUST be at the top.
 
 require_once 'admin/_main.php';         // Admin Settings.
-require_once 'ways_to_watch/_main.php'; // Ways to Watch.
 require_once 'assets/symbolicons.php';  // Symbolicons/Font Icons.
 require_once 'blocks/_main.php';        // Custom Blocks.
+require_once 'debugger/_main.php';      // Debugger and data fixers.
 require_once 'plugins/_main.php';       // Tweaks for Plugins.
 require_once 'rest-api/_main.php';      // Our Rest API.
 require_once 'statistics/_main.php';    // Stats.
 require_once 'this-year/_main.php';     // This Year.
+require_once 'ways_to_watch/_main.php'; // Ways to Watch.
 
 require_once 'cpts/_main.php';         // Custom Post Types: This MUST be at the end.
